+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/LectureseriesLocalService.java
@@ -255,6 +255,6 @@ public interface LectureseriesLocalService extends BaseLocalService,
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public java.util.List<de.uhh.l2g.plugins.model.Lectureseries> getFilteredBySemesterFacultyProducer(
-		java.lang.Integer approved, java.lang.String semester, int facultyId,
-		int producerId);
+		java.lang.Integer approved, java.lang.String semester,
+		java.lang.Long facultyId, java.lang.Long producerId);
 }
