+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/LectureseriesLocalServiceUtil.java
@@ -281,8 +281,8 @@ public class LectureseriesLocalServiceUtil {
 	}
 
 	public static java.util.List<de.uhh.l2g.plugins.model.Lectureseries> getFilteredBySemesterFacultyProducer(
-		java.lang.Integer approved, java.lang.String semester, int facultyId,
-		int producerId) {
+		java.lang.Integer approved, java.lang.String semester,
+		java.lang.Long facultyId, java.lang.Long producerId) {
 		return getService()
 				   .getFilteredBySemesterFacultyProducer(approved, semester,
 			facultyId, producerId);
