+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/LectureseriesLocalServiceWrapper.java
@@ -296,8 +296,8 @@ public class LectureseriesLocalServiceWrapper
 
 	@Override
 	public java.util.List<de.uhh.l2g.plugins.model.Lectureseries> getFilteredBySemesterFacultyProducer(
-		java.lang.Integer approved, java.lang.String semester, int facultyId,
-		int producerId) {
+		java.lang.Integer approved, java.lang.String semester,
+		java.lang.Long facultyId, java.lang.Long producerId) {
 		return _lectureseriesLocalService.getFilteredBySemesterFacultyProducer(approved,
 			semester, facultyId, producerId);
 	}
