+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/persistence/LectureseriesFinder.java
@@ -21,6 +21,6 @@ public interface LectureseriesFinder {
 	public java.util.List<java.lang.String> findAllSemesters(int begin, int end);
 
 	public java.util.List<de.uhh.l2g.plugins.model.Lectureseries> findeFilteredByApprovedSemesterFacultyProducer(
-		java.lang.Integer approved, java.lang.String semester, int facultyId,
-		int producerId);
+		java.lang.Integer approved, java.lang.String semester,
+		java.lang.Long facultyId, java.lang.Long producerId);
 }
