+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/persistence/LectureseriesFinderUtil.java
@@ -27,8 +27,8 @@ public class LectureseriesFinderUtil {
 	}
 
 	public static java.util.List<de.uhh.l2g.plugins.model.Lectureseries> findeFilteredByApprovedSemesterFacultyProducer(
-		java.lang.Integer approved, java.lang.String semester, int facultyId,
-		int producerId) {
+		java.lang.Integer approved, java.lang.String semester,
+		java.lang.Long facultyId, java.lang.Long producerId) {
 		return getFinder()
 				   .findeFilteredByApprovedSemesterFacultyProducer(approved,
 			semester, facultyId, producerId);
